METAMAC-1836: [API] Cambiar tipo de dato xsd:integer por xsd:int
git-svn-id: http://svn.arte-consultores.com/svn/arte/istac/metamac/metamac-srm/trunk/metamac-srm@48251 05f23f4b-1f0b-0410-ba1b-ad9509270e59

+++ metamac-srm-rest-internal-impl/src/test/java/org/siemac/metamac/srm/rest/internal/v1_0/concept/mapper/ConceptsDo2RestMapperTest.java
@@ -301,11 +301,11 @@ public class ConceptsDo2RestMapperTest {
         QuantityChangeRate quantityChangeRate = (QuantityChangeRate) target.getQuantity();
         assertEquals("urn:sdmx:org.sdmx.infomodel.codelist.Code=agency01:codelist01(01.000).unitCode01", quantityChangeRate.getUnitCode().getUrn());
         assertEquals(QuantityUnitSymbolPosition.START, quantityChangeRate.getUnitSymbolPosition());
-        assertEquals(BigInteger.valueOf(5), quantityChangeRate.getSignificantDigits());
-        assertEquals(BigInteger.valueOf(2), quantityChangeRate.getDecimalPlaces());
-        assertEquals(BigInteger.valueOf(10), quantityChangeRate.getUnitMultiplier());
-        assertEquals(BigInteger.valueOf(1000), quantityChangeRate.getMin());
-        assertEquals(BigInteger.valueOf(2000), quantityChangeRate.getMax());
+        assertEquals(Integer.valueOf(5), quantityChangeRate.getSignificantDigits());
+        assertEquals(Integer.valueOf(2), quantityChangeRate.getDecimalPlaces());
+        assertEquals(10, quantityChangeRate.getUnitMultiplier());
+        assertEquals(Integer.valueOf(1000), quantityChangeRate.getMin());
+        assertEquals(Integer.valueOf(2000), quantityChangeRate.getMax());
         assertEquals("urn:sdmx:org.sdmx.infomodel.conceptscheme.Concept=agency01:conceptScheme02(01.000).numerator1", quantityChangeRate.getNumerator().getUrn());
         assertEquals("urn:sdmx:org.sdmx.infomodel.conceptscheme.Concept=agency01:conceptScheme03(01.000).denominator1", quantityChangeRate.getDenominator().getUrn());
         assertEquals(false, quantityChangeRate.isIsPercentage());
